web_demo/services/inference/pyproject.toml
[project]
name = "inference-service"
version = "0.1.0"
description = "MAE + U-Net inpainting + classifier inference service and CLI."
requires-python = ">=3.10"
dependencies = [
    "torch>=2.2",
    "torchvision>=0.17",
    "transformers>=4.40",
    "pillow>=10.0",
    "fastapi>=0.110",
    "uvicorn[standard]>=0.29",
    "pydantic>=2.6",
]

[project.scripts]
inference-cli = "inference_service.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/inference_service"]
